+++ lib/state_management/core/reactive_notifier.dart
@@ -1,7 +1,7 @@
 
 
-class ReactiveNotifier<T> {
-  final Stream<T> stream;
-
-  ReactiveNotifier(this.stream);
-}
+// class ReactiveNotifier<T> {
+//   final Stream<T> stream;
+//
+//   ReactiveNotifier(this.stream);
+// }